==> R/04 REPORT Demographic.community.spread.Rmd <==
---
title: Postcode characteristics of Melbourne's COVID19 hotspots
author: Monika Sarder
date: '2020-09-11'
categories:
  - COVID19
tags:
  - COVID19
  - Melbourne
  - Melbourne lockdown
  - JobSeeker
draft: no  
---

```{r, include = FALSE}
knitr::opts_chunk$set(collapse = TRUE, out.width = "100%")

library(rmarkdown)
library(plotly)
library(forcats)
library(knitr)
library(kableExtra)
library(tidyverse)
library(DT)
library(RColorBrewer)
library(downloadthis)
library(sf)
library(viridis)
all <- readRDS("../Outputs/03.Melbourne.spatial.COVID19.RDS")

exclusion.list <- all %>%
  st_set_geometry(NULL) %>%
  filter(Include == "No") %>%
  select(Postcode, Suburb, Population)

write.csv(exclusion.list, "../exclusion.list.csv", row.names = FALSE)
```

Melbourne’s second wave, which commenced in June and peaked in the first week of August, was driven primarily by community transmission. New outbreaks throughout this period have tended to arise in workplaces, flowing through to families and communities (Source: [DHHS](https://www.dhhs.vic.gov.au/coronavirus-update-victoria-07-september-2020)).   

The highly localised nature of the second wave led to experimentation with a Postcode-based lockdown in July. This was quickly abandoned in favour of a city-wide lockdown. However, it is clear that geography, and geospatial characteristics, have an effect on the [rate of virus transmission](https://www.covid19data.com.au/victoria).  

This social and economic reality was notably absent from the [modelling](https://www.mja.com.au/journal/2020/probability-6-week-lockdown-victoria-commencing-9-july-2020-achieving-elimination) that informed Melbourne’s roadmap out of lockdown. The model assumed identical features across postcodes, that all cases of COVID19 in the community were randomly distributed, and that all cases are mixing in the community in the same way.   

In order to achieve a successful COVID19 endgame, we need to identify the actual community characteristics that are correlated with high rates of transmission. A detailed understanding of the reality of the city, has informational value relevant to both the modelling, and the public health response.

The final dataset comprises 222 Melbourne postcodes.^[Twenty-four of the postcodes were removed from the dataset because the population was too small to be meaningful (<1500 residents), or because an area had undergone a major redevelopment since the Census was undertaken, and the data no longer reflected the population. Excluded postcodes and reasons for exclusion are listed `r xfun::embed_files('../Outputs/exclusion.list.csv', text = "here")`] 

# Data in this study

The [ABS 2016 Census of Population and Housing](https://datapacks.censusdata.abs.gov.au/geopackages/) contains rich information on a range of community characteristics. Statistical analysis was undertaken of the relationship between key features and the level of COVID19 infection in various postcodes (ie number of confirmed cases per 100K) at the peak of the wave on 6 August.

The following characteristics showed positive and statistically significant relationships between COVID19 infection levels and the proportion of residents:  

*  working in the transport industry; and  
*  who are below the median age
*  living below the poverty line  
*  living in rental properties  
*  living in a flat or semi-detached home (cf a detached house) 
*  living in a home without an internet connection  
*  speaking a language other than English in the home  
 
**A note on interpretation**

A positive statistical correlation indicates that when we see an increase in one characteristic (eg number of rental households), we can expect the other characteristic (ie COVID19 cases) to increase as well.  An important caveat to readers, is that such a relationship does not constitute ‘proof’ that one thing causes another, merely that something connects them. 

For example, the number of ice creams sold per day may be positively correlated with the number of beach lifesaver rescues. This does not mean however, that ice-creams are causing people to get into trouble in the water.

Statisticians deal with the relationships, or patterns, between observable and measurable phenomena. When you are dealing with the statistics of human behaviour (ie in this case in relation to a public health crisis and highly transmissible virus), there is always more to the story, more questions to ask, more to uncover. Statistics cannot provide you with a perfect answer, but it can show you where to look to uncover a better answer than the one you already have. 

# Geographical concentration of COVID19

At the peak of the second wave the number of confirmed cases per 100K of population was significantly higher in Melbourne’s North and West than in other parts of the city.

**Figure 1 Confirmed cases per 100K by Postcode on 6 August**

```{r choropleth creation, echo = FALSE, warning = FALSE}
knitr::opts_chunk$set(collapse = TRUE, out.width = "100%")

vis <- all %>%
  select(Postcode, Suburb, Population, Cases.per.100K, Confirmed, Include) %>%
  mutate(Cases.per.100K = ifelse(Include == "No", 0, Cases.per.100K))

g <- list(
  showlegend = FALSE,
  showframe = FALSE,
  showcoastlines = FALSE,
  projection = list(type = "Mercator")
)

geospatial.plot <- vis %>%
  plot_geo(
    split = ~Postcode, showlegend = FALSE, hoverinfo = "text", color = ~Cases.per.100K,
    colors = viridis_pal(option = "B")(30),
    text = ~ paste(
      "Postcode:", Postcode,
      "<br>", "Suburb:", Suburb,
      "<br>", "Cases per 100K:", Cases.per.100K,
      "<br>", "All cases 6 Aug:", Confirmed
    )
  ) %>%
  add_sf() %>%
  layout(geo = g)

colorbar(geospatial.plot, len = 0.4, title = "Cases per 100K 6 Aug")
```

<font size="2">**Sources: [DHHS](https://www.theage.com.au/national/victoria/victoria-coronavirus-data-find-out-how-active-case-numbers-have-changed-in-your-postcode-in-the-past-week-20200806-p55j9n.html); 2016 Census - [Selected Medians and Averages; Industry of Employment by Occupation](https://datapacks.censusdata.abs.gov.au/geopackages/)**</font>

A full list of Melbourne postcodes, COVID19 cases and Population data appears below.

**Table 1 COVID19 confirmed cases 6 August and population by Postcode**

```{r list function, echo = FALSE, warning = FALSE}
create_dt <- function(x, pl) {
  DT::datatable(x,
    rownames = FALSE,
    extensions = c("Buttons"),
    options = list(
      scrollX = TRUE, pageLength = pl, dom = "Blfrtip",
      buttons = c("copy", "csv", "excel")
    )
  )
}
```


```{r list data, echo = FALSE, warning = FALSE}
all <- all %>% st_set_geometry(NULL)

main.table <- all %>%
  select(Postcode, Suburb, Confirmed, Population, Cases.per.100K, Include) %>%
  arrange(desc(Confirmed)) %>%
  rename(
    `Confirmed 6 Aug` = Confirmed,
    Pop = Population,
    `Cases per 100K` = Cases.per.100K,
    Included = Include
  )

create_dt(main.table, pl = 10)
```

```{r filter excluded, echo = FALSE, warning = FALSE}
dat <- all %>% filter(Include == "Yes")
```

# COVID19 and resident activities  



```{r age and transport, echo = FALSE, warning = FALSE}
age.cor <- cor.test(dat$Median.age, dat$Cases.per.100K)

transport.cor <- cor.test(dat$Employed.transport.Percent, dat$Cases.per.100K)
```

Several academics have called for a Postcode-based easing of restrictions, referencing [only disadvantage factors](https://theconversation.com/open-covid-cold-spots-first-a-way-out-of-lockdown-for-melbourne-145387).

This kind of a analysis ignores key variables. A more holistic approach involves looking at what community members actually do, and how this predicates movement.

Occupation data reveals that a relatively high proportion of community members in COVID19 hotspots work in the transport industry. It should come as no surprise that communities that are home to individuals involved in moving freight, goods, warehousing and delivering shopping and takeaway, should be linked to COVID19 risk. The correlation of the percentage of employed persons working in transports and COVID19 is quite high, at `r round(transport.cor$estimate, digits = 3)` (p-value = <0.000). 

In terms of demographics, the median age of community members is also strongly, and significantly, correlated with COVID19, at `r round(age.cor$estimate, digits = 3)` (p-value = <0.000). That is, communities that skew younger have higher rates of infection. This makes sense, as younger community members are more likely to have multiple responsibilities, such as working, caring for children, caring for parents etc. These responsibilities involve both cross-city movement and interpersonal contact. 

Premier Daniel Andrews rightly and sensibly rejected a return to Postcode-based approaches at the start of September, stating that essential ["workers live in families"](https://www.youtube.com/watch?v=NBiqRkW_q6k&t=1s&ab_channel=ABCNews%28Australia%29), need to travel and therefore, functional isolation is not possible.

**Figure 5 COVID19 against age and work - by Postcode**

```{r age and transport vis, echo = FALSE, warning = FALSE}
fit.age <- lm(Cases.per.100K ~ Median.age, data = dat)
p1 <- dat %>%
  plot_ly(
    x = ~Median.age, y = ~Cases.per.100K,
    hoverinfo = "text",
    text = ~ paste(
      "Postcode:", Postcode, "<br>",
      "Suburb:", Suburb, "<br>",
      "Cases:", Cases.per.100K, "<br>",
      "Median age:", Median.age
    )
  ) %>%
  add_markers(name = ~"Median age") %>%
  add_lines(x = ~Median.age, y = ~ fitted(fit.age), showlegend = F)

fit.transport <- lm(Cases.per.100K ~ Employed.transport.Percent, data = dat)

p2 <- dat %>%
  plot_ly(
    x = ~Employed.transport.Percent, y = ~Cases.per.100K,
    hoverinfo = "text",
    text = ~ paste(
      "Postcode:", Postcode, "<br>",
      "Suburb:", Suburb, "<br>",
      "Cases:", Cases.per.100K, "<br>",
      "(%) Employed transport:", Employed.transport.Percent
    )
  ) %>%
  add_markers(name = ~"(%) Employed transport") %>%
  add_lines(x = ~Employed.transport.Percent, y = ~ fitted(fit.transport), showlegend = F)

plotly::subplot(p2, p1, nrows = 1) %>%
  layout(
    yaxis = list(title = "Cases per 100K", range = c(0, 1790)),
    yaxis2 = list(range = c(0, 1790)),
    xaxis2 = list(title = "Median age"),
    xaxis = list(title = "(%) Transport")
  )
```

<font size="2">**Source: 2016 Census - [Selected Medians and Averages; Industry of Employment by Occupation](https://datapacks.censusdata.abs.gov.au/geopackages/)**</font>  

# COVID19 and security of housing and income support  

```{r housing and poverty, echo = FALSE, warning = FALSE}
poverty.cor <- cor.test(dat$Low.income.Percent, dat$Cases.per.100K)

rental.cor <- cor.test(dat$Rented.Percent, dat$Cases.per.100K)
```

A positive correlation was identified between the percentage of residents whose personal income is less than \$500 per week and COVID19 cases, with a correlation of `r round(poverty.cor$estimate, digits = 3)` (p-value = <0.000).  

This correlation is deeply concerning, given that Federal JobSeeker and JobKeeper rates are due to fall dramatically later this month, pushing more than [100,000 Victorians below the poverty line](https://www.tai.org.au/content/jobseeker-cut-push-370000-poverty-including-80000-children) (defined at [$457](https://newsroom.unsw.edu.au/news/social-affairs/unsw-and-acoss-report-shows-3m-australians-living-poverty#:~:text=The%202020%20Poverty%20in%20Australia,week%20for%20a%20single%20adult.) per week). Coupled with the lack of Paid Pandemic Leave for all workers, increased financial pressure on essential workers to make ends meet, could undermine efforts to stop the spread.

A positive correlation was also identified between the percentage of Postcode residents who lived in rental properties and the level of COVID19 cases, with a correlation of `r round(rental.cor$estimate, digits = 3)` (p-value = `r round(rental.cor$p.value, digits = 3)`).   

This suggests that having security of tenure – the confidence that you will not be kicked out of your home – may be relevant to public health.   

Policy measures such as eviction moratoriums, rent relief, and longer term investments in affordable housing and public housing, likely have an important part to play in achieving public health objectives. 

**Figure 2 COVID19 against rent, income - by Postcode**

```{r housing and poverty vis, echo = FALSE, warnings = FALSE}
fit.perc <- lm(Cases.per.100K ~ Low.income.Percent, data = dat)

p1 <- dat %>%
  plot_ly(
    x = ~Low.income.Percent, y = ~Cases.per.100K,
    hoverinfo = "text",
    text = ~ paste(
      "Postcode:", Postcode, "<br>",
      "Suburb:", Suburb, "<br>",
      "Cases:", Cases.per.100K, "<br>",
      "Percentage under 500:", Low.income.Percent
    )
  ) %>%
  add_markers(name = ~"(%) Under $500 ") %>%
  add_lines(x = ~Low.income.Percent, y = ~ fitted(fit.perc), showlegend = F)

fit.rent <- lm(Cases.per.100K ~ Rented.Percent, data = dat)

p2 <- dat %>%
  plot_ly(
    x = ~Rented.Percent, y = ~Cases.per.100K,
    hoverinfo = "text",
    text = ~ paste(
      "Postcode:", Postcode, "<br>",
      "Suburb:", Suburb, "<br>",
      "Cases:", Cases.per.100K, "<br>",
      "(%) Rental:", Rented.Percent
    )
  ) %>%
  add_markers(name = ~"(%) Rental") %>%
  add_lines(x = ~Rented.Percent, y = ~ fitted(fit.rent), showlegend = F)

plotly::subplot(p1, p2, nrows = 1) %>%
  layout(
    yaxis = list(title = "Cases per 100K"),
    xaxis = list(title = "(%) Under $500"),
    xaxis2 = list(title = "(%) Rental")
  )
```
 <font size="2"> **Source: 2016 Census - [Total Personal Income (Weekly) by Age by Sex; Tenure Type and Landlord Type by Dwelling Structure](https://datapacks.censusdata.abs.gov.au/geopackages/)**</font>

# COVID19 and dwelling structure 

The relationship between dense living conditions and spread was thrown into sharp relief in July, [when nine towers in Flemington and North Melbourne were placed under hard lockdown following COVID-19 outbreaks](https://www.abc.net.au/news/2020-07-17/melbourne-victoria-public-housing-social-mix-redevelopment/12459870).

Public housing has subsequently been a focus of the fight against COVID19. The Government recently announced [a new policy which will make alternative rental properties available to people in towers, provided that they meet specific risk criteria](https://www.theage.com.au/national/victoria/public-tower-tenants-offered-relocation-to-fight-spread-of-covid-19-20200829-p55qip.html).  

Table 2 below, shows the relevant relationships for percentage of residents living in different dwelling types, and COVID19. There is a negative relationship between the percentage of residents who live in detached houses and COVID19. Meanwhile, there is a positive relationship with the percentage of residents who live in semi-detached dwellings, flats or apartments. 

This has implications for how the State approaches public health messaging to residents of higher density dwellings, particularly around shared spaces and communal surfaces. In addition, 'pandemic proofing' may be a consideration informing future housing design.

**Table 2 Correlation of COVID19 and Dwelling structure**
```{r dwellin structure, warning = FALSE, echo = FALSE}
det.cor <- cor.test(dat$House.Percent, dat$Cases.per.100K)
semi.det.cor <- cor.test(dat$Semi.detached.Percent, dat$Cases.per.100K)
apartment.cor <- cor.test(dat$Flat.Percent, dat$Cases.per.100K)

correlation.main.dwellings <- data.frame(
    Dwelling_type = c("Detached house", "Semi-detached", "Apartment or flat"),
    Correlation = c(det.cor$estimate, semi.det.cor$estimate, apartment.cor$estimate),
    P_value = c(det.cor$p.value, semi.det.cor$p.value, apartment.cor$p.value)
  ) %>%
  mutate(
    Correlation = round(Correlation, 3),
    P_value = round(P_value, 3)
  )

kable(correlation.main.dwellings) %>%
  kable_styling(full_width = F, position = "left") %>%
  row_spec(1, color = "blue") %>%
  row_spec(2:3, color = "red")
```

**Figure 3 COVID19 against housing type - by Postcode**
```{r dwelling structure vis, echo = FALSE, warning = FALSE}
fit.det <- lm(Cases.per.100K ~ Semi.detached.Percent, data = dat)

p1 <- dat %>%
  plot_ly(
    x = ~Semi.detached.Percent, y = ~Cases.per.100K,
    hoverinfo = "text",
    text = ~ paste(
      "Postcode:", Postcode, "<br>",
      "Suburb:", Suburb, "<br>",
      "Cases:", Cases.per.100K, "<br>",
      "(%) Semi-detached:", Semi.detached.Percent
    )
  ) %>%
  add_markers(showlegend = FALSE) %>%
  add_lines(x = ~Semi.detached.Percent, y = ~ fitted(fit.det), showlegend = FALSE)

fit.house <- lm(Cases.per.100K ~ House.Percent, data = dat)

p2 <- dat %>%
  plot_ly(
    x = ~House.Percent, y = ~Cases.per.100K,
    hoverinfo = "text",
    text = ~ paste(
      "Postcode:", Postcode, "<br>",
      "Suburb:", Suburb, "<br>",
      "Cases:", Cases.per.100K, "<br>",
      "(%) House:", House.Percent
    )
  ) %>%
  add_markers(showlegend = FALSE) %>%
  add_lines(x = ~House.Percent, y = ~ fitted(fit.house), showlegend = FALSE)

fit.flat <- lm(Cases.per.100K ~ Flat.Percent, data = dat)

p3 <- dat %>%
  plot_ly(
    x = ~Flat.Percent, y = ~Cases.per.100K,
    hoverinfo = "text",
    text = ~ paste(
      "Postcode:", Postcode, "<br>",
      "Suburb:", Suburb, "<br>",
      "Cases:", Cases.per.100K, "<br>",
      "(%) Flat.Percent:", Flat.Percent
    )
  ) %>%
  add_markers(showlegend = FALSE) %>%
  add_lines(x = ~Flat.Percent, y = ~ fitted(fit.flat), showlegend = FALSE)

plotly::subplot(p2, p1, p3, nrows = 1) %>%
  layout(
    yaxis = list(title = "Cases per 100K"),
    xaxis = list(title = "(%) House"),
    xaxis2 = list(title = "(%) Semi-detached"),
    xaxis3 = list(title = "(%) Flat")
  )
```

<font size="2">**2016 Census - [Dwelling Structure](https://datapacks.censusdata.abs.gov.au/geopackages/)**</font>

# COVID19 and timely access to public health information

```{r language and internet, echo = FALSE, warning = FALSE}
language.cor <- cor.test(dat$Lang.NE.Percent, dat$Cases.per.100K)
no.internet.cor <- cor.test(dat$No.internet.Percent, dat$Cases.per.100K)
```

Given the highly dynamic nature of the Victorian public health crisis, timely access to public health messages is absolutely critical.

An initial lack of engagement with migrant communities was acknowledged to be a problem by the Victorian Government. In July, Chief Health Officer Brett Sutton said:

["We know that there are some migrant communities, recent migrants or culturally and linguistically diverse communities, who are over-represented now with some of our new cases […] It's our obligation as government to reach those people. It's not their fault if we're not going in with appropriate engagement."](https://www.abc.net.au/news/2020-06-29/coronavirus-multilingual-australia-missing-out-covid-19-info/12403510 )

At the peak of the second wave, there was a positive correlation between the number of Postcode residents who did not speak English at home and infection numbers, measured at `r round(language.cor$estimate, digits = 3)` (p-value = <0.000). Significant resources have now been directed towards resource translation, as well as going door to door.

There was also positive correlation of `r round(no.internet.cor$estimate, digits = 3)` (p-value = <0.000), between the proportion of households with no internet access, and COVID19. The digital divide may be posing a significant barrier to compliance with safety measures.

**Figure 4 COVID19 against Home language, Internet access - by Postcode**
```{r language and internet vis, echo = FALSE, warnings = FALSE}
fit.lang <- lm(Cases.per.100K ~ Lang.NE.Percent, data = dat)
p1 <- dat %>%
  plot_ly(
    x = ~Lang.NE.Percent, y = ~Cases.per.100K,
    hoverinfo = "text",
    text = ~ paste(
      "Postcode:", Postcode, "<br>",
      "Suburb:", Suburb, "<br>",
      "Cases:", Cases.per.100K, "<br>",
      "Home language not English:", Lang.NE.Percent
    )
  ) %>%
  add_markers(name = ~"(%) Not English ") %>%
  add_lines(x = ~Lang.NE.Percent, y = ~ fitted(fit.lang), showlegend = F)

fit.internet <- lm(Cases.per.100K ~ No.internet.Percent, data = dat)

p2 <- dat %>%
  plot_ly(
    x = ~No.internet.Percent, y = ~Cases.per.100K,
    hoverinfo = "text",
    text = ~ paste(
      "Postcode:", Postcode, "<br>",
      "Suburb:", Suburb, "<br>",
      "Cases:", Cases.per.100K, "<br>",
      "(%) No internet households:", No.internet.Percent
    )
  ) %>%
  add_markers(name = ~"(%) Households no internet") %>%
  add_lines(x = ~No.internet.Percent, y = ~ fitted(fit.internet), showlegend = F)

plotly::subplot(p1, p2, nrows = 1) %>%
  layout(
    yaxis = list(title = "Cases per 100K"),
    xaxis = list(title = "(%) Non English"),
    xaxis2 = list(title = "(%) No internet")
  )
```
 <font size="2"> **2016 Census - [Language Spoken at Home by Proficiency in Spoken English/Language by Sex; Dwelling Internet Connection by Dwelling Structure](https://datapacks.censusdata.abs.gov.au/geopackages/)**</font>
 
# Conclusion

Statisticians are often encouraged to reduce differences down to a single disadvantage score or measure, because it is an elegant and concise way to explain phenomena. 

However when it comes to public health, oversimplifying the characteristics of 'disadvantaged' suburbs, can lead to adverse outcomes. This is because different community characteristics  call for different policy solutions to stop the spread.

The association with renting, for example, highlights the importance of eviction moratoriums and rent relief, whereas the association with language spoken at home, foregrounds the importance of linguistically appropriate resources.

The prevalence of transport work is also an important characteristic, since movement of essential freight and logistics workers is not something that we can, or would want to curtail.  

#  Data downloads 

```{r data downloads, echo = FALSE, warnings = FALSE}
# make sure exclusions are noted
exclude <- all %>%
  filter(
    Population < 1500 | 
    Suburb %in% c("Plenty", "Somerton", "Ardeer, Deer Park East") |
    Postcode %in% c("3062", "3010")
  ) %>%
  pull(Postcode)

all <- all %>%
  mutate(Exclude = ifelse(Postcode %in% exclude, "No", "Yes"))

standard <- c("Postcode", "Include", "Confirmed", "Population", "Cases.per.100K")

transport <- all %>%
  select(all_of(standard), Employed.transport, Employed.Persons, Employed.transport.Percent)

write.csv(transport, "../Outputs/transport.csv", row.names = FALSE)

xfun::embed_files("../Outputs/transport.csv", text = "Download Transport sector employment and COVID19 data as csv")
```
\
```{r age data, echo = FALSE, warnings = FALSE}
age <- all %>%
  select(all_of(standard), Median.age, Median.household.income)

write.csv(age, "../Outputs/age.csv", row.names = FALSE)

xfun::embed_files("../Outputs/age.csv", text = "Download Median age and COVID19 data as csv")
```
\
```{r income data, echo = FALSE, warnings = FALSE}
low.income <- all %>%
  select(all_of(standard), Low.income.Persons, Total.over.15, Low.income.Percent)

write.csv(low.income, "../Outputs/low.income.csv", row.names = FALSE)

xfun::embed_files("../Outputs/low.income.csv", text = "Download Low weekly income and COVID19 data as csv")
```
\
```{r rent data, echo = FALSE, warnings = FALSE}
rental.tenure <- all %>%
  select(all_of(standard), Rented.Dwellings, Occupied.dwellings, Rented.Percent)

write.csv(rental.tenure, "../Outputs/renting.csv", row.names = FALSE)

xfun::embed_files("../Outputs/renting.csv", text = "Download rental tenure and COVID19 data as csv")
```
\
```{r transport data, echo = FALSE, warnings = FALSE}
dwelling <- all %>%
  select(all_of(standard), Semi.detached.Persons, Flat.Persons, House.Persons, Dwelling.persons, Semi.detached.Percent, Flat.Percent, House.Percent)

write.csv(dwelling, "../Outputs/dwelling.csv", row.names = FALSE)

xfun::embed_files("../Outputs/dwelling.csv", text = "Download Dwelling type Person level data as csv")
```
\
```{r language data, echo = FALSE, warnings = FALSE}
non.english <- all %>%
  select(all_of(standard), Language.NE, Population, Lang.NE.Percent)

write.csv(age, "../Outputs/language.csv", row.names = FALSE)

xfun::embed_files("../Outputs/language.csv", text = "Language other than English and COVID19 data as csv")
```
\
```{r internet data, echo = FALSE, warnings = FALSE}
non.internet <- all %>%
  select(all_of(standard), No.home.access, Occupied.dwellings, No.internet.Percent)

write.csv(age, "../Outputs/internet.csv", row.names = FALSE)

xfun::embed_files("../Outputs/internet.csv", text = "Home internet access and COVID19 data as csv")
```
